feat(sandbox): Phase 2 — BusyBox, SandboxClient, IRQ priority fix
Add BusyBox static to initramfs (auto-download from Alpine APK),
providing sh/cat/ls/grep/diff and 50+ shell tools inside the guest VM.

Fix virtio interrupt starvation: vtimer fired at HZ rate and always
won ICC_IAR1_EL1 priority over virtio-console, preventing multi-exec.
Reorder IRQ dispatch to service virtio devices before vtimer.

Add SandboxClient high-level API wrapping SandboxPool with
exec/read_file/write_file/git_diff/apply_to_host for Worker integration.

Additional: settle() after exec, VmSnapshot::invalidate_cache(),
ExecResult Debug derive, clippy fixes, remove all #[ignore].

+++ crates/opengoose-sandbox/src/fuse/mod.rs
@@ -388,7 +388,7 @@ mod tests {
         let resp = build_error_response(42, libc::ENOENT);
         let header: FuseOutHeader = unsafe { std::ptr::read_unaligned(resp.as_ptr() as *const _) };
         assert_eq!(header.len, FUSE_OUT_HEADER_SIZE as u32);
-        assert_eq!(header.error, -(libc::ENOENT as i32));
+        assert_eq!(header.error, -libc::ENOENT);
         assert_eq!(header.unique, 42);
     }
